feat: switch project from ts to js
