--- .eslintignore.internal
@@ -1,2 +1,0 @@
-/src/__tests__/bad/
-/src/__tests__/good/
