+++ .eslintrc.js
@@ -1,3 +1,3 @@
 module.exports = {
-  extends: ['./dist/index.js'],
+  extends: ['./src/index.js'],
 };
