+++ .prettierrc.js
@@ -1,3 +1,3 @@
 module.exports = {
-  ...require('./dist/prettier'),
+  ...require('./src/prettier'),
 };
