+++ package.json
@@ -7,11 +7,11 @@
   },
   "scripts": {
     "version": "npm run build",
-    "build": "rimraf dist && tsc -p tsconfig.build.json --outDir dist",
-    "lint": "npm run build && eslint --ext ts,js,tsx,jsx,json ./src --ignore-path ./.eslintignore.internal",
+    "build": "",
+    "lint": "eslint --ext ts,js,tsx,jsx,json ./src --ignore-path ./src/__tests__/.eslintignore",
     "format": "npm run lint -- --fix",
     "typecheck": "tsc --noEmit",
-    "test": "npm run build && node dist/__tests__/test.js"
+    "test": "node src/__tests__/test.js"
   },
   "author": "SMG Automotive",
   "license": "MIT",
@@ -38,8 +38,10 @@
     "eslint-plugin-testing-library": "5.3.1"
   },
   "devDependencies": {
-    "@types/node": "17.0.31",
-    "@types/prettier": "2.6.0",
+    "@rollup/plugin-commonjs": "22.0.0",
+    "@rollup/plugin-multi-entry": "4.1.0",
+    "@rollup/plugin-node-resolve": "13.3.0",
+    "@rollup/plugin-typescript": "8.3.2",
     "chalk": "4.1.2",
     "jest": "28.0.3",
     "prettier": "2.6.2",
@@ -47,17 +49,15 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "rimraf": "3.0.2",
-    "semantic-release": "19.0.2",
-    "ts-node": "10.7.0",
-    "typescript": "4.6.4"
+    "semantic-release": "19.0.2"
   },
   "exports": {
-    "./prettier": "./dist/prettier.js",
-    "./next": "./dist/next.js",
-    "./react": "./dist/react.js",
-    ".": "./dist/index.js"
+    "./prettier": "./src/prettier.js",
+    "./next": "./src/next.js",
+    "./react": "./src/react.js",
+    ".": "./src/index.js"
   },
   "files": [
-    "dist/**/*"
+    "src/**/*"
   ]
 }
