+++ src/__tests__/.eslintignore
@@ -1,0 +1,1 @@
+/src/__tests__/
