+++ src/__tests__/.eslintrc.next.js
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
 module.exports = {
-  extends: ['../../dist/next.js'],
+  extends: ['../../src/next.js'],
   rules: {
     '@next/next/no-html-link-for-pages': 'off',
   },
