+++ src/__tests__/bad/js/internal-modules.js
@@ -1,2 +1,2 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+// eslint-disable-next-line no-unused-vars
 import picka from '@next/eslint-plugin-next/lib/index';
