+++ src/__tests__/bad/js/no-unresolved.js
@@ -1,2 +1,2 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+// eslint-disable-next-line no-unused-vars
 import picka from '@pika/pack/dist-src/index';
