+++ src/__tests__/bad/ts/no-unresolved.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import options from 'eslint/lib/options';
