+++ src/__tests__/good/js/semicolons.js
@@ -1,1 +1,2 @@
+// eslint-disable-next-line no-unused-vars
 const _yo = true;
