--- tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": [
-    "dist",
-    "node_modules",
-    "src/__tests__/good/",
-    "src/__tests__/bad/"
-  ],
-  "include": ["src"]
-}
